# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: gagebu
  description: SU - 본인의 소비내역을 기록/관리 하는 가계부 API 서버
  version: 0.0.1
servers:
- url: http://localhost:8080
...
paths:
...
  /api/v1/users:
    post:
      tags:
      - User API
      summary: 회원가입 API
      description: 회원가입을 요청하는 API 입니다.
      operationId: user-signup
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/회원가입 요청'
            examples:
              user-signup:
                value: "{\r\n  \"email\" : \"c24154c6ad@test.com\",\r\n  \"password\"\
                  \ : \"0b7255f5-6121-4279-99cc-4ddbd64ee3f7\"\r\n}"
      responses:
        "201":
          description: "201"
          headers:
            Location:
              description: 회원가입 된 유저 URI
              schema:
                type: string
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/회원가입 응답'
              examples:
                user-signup:
                  value: "{\r\n  \"id\" : 15\r\n}"
...
  /api/v1/users/{userId}:
    get:
      tags:
      - User API
      summary: 유저 조회 API
      description: 유저를 단일 조회하는 API 입니다.
      operationId: user-findOne
      parameters:
      - name: userId
        in: path
        description: 유저 아이디
        required: true
        schema:
          type: string
      - name: Authorization
        in: header
        description: JWT accessToken
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJIUzUxMiJ9.eyJyb2xlIjoiTE9HSU4iLCJ1c2VySWQiOjE2LCJpYXQiOjE2NjQzMDQ1ODksImV4cCI6MTY2NDMwNjM4OX0.KPpT6Z5ZDUf4hJgSSg3cuMYiESJNATd8YjMC6WqeMaVmVBZi22H5zvur6TVmaJmhv7Y6m8938oWXafhbVwaoAA
      responses:
        "200":
          description: "200"
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/유저 단일 조회 응답'
              examples:
                user-findOne:
                  value: "{\r\n  \"userId\" : 16,\r\n  \"email\" : \"c52812fab2@test.com\"\
                    \r\n}"
      security:
      - bearerAuthJWT: []
components:
...
  securitySchemes:
    bearerAuthJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT